+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@ SwiftUIKit makes its best effort to honor semver, but breaking changes can occur
 ### 🗑️ Deprecations
 
 * `AlertContext` has been moved to https://github.com/danielsaidi/PresentationKit.
+* `ErrorAlerter` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `FullScreenCoverContext` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `Image+ScanCode` has been moved to https://github.com/danielsaidi/ScanCodes.
 * `LinkText` has been moved to https://github.com/danielsaidi/TextReplacements.
@@ -22,6 +23,7 @@ SwiftUIKit makes its best effort to honor semver, but breaking changes can occur
 * `PresentationContext` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `ScanCodeType` has been moved to https://github.com/danielsaidi/ScanCodes.
 * `SheetContext` has been moved to https://github.com/danielsaidi/PresentationKit.
+* `UrlSharing` has been deprecated.